requirements.txt
fastapi
uvicorn
jinja2
websockets
python-dotenv
sqlalchemy
python-multipart
pydantic
PyJWT
gunicorn
supabase